Improve Oban plugin handler and fix error handling
The main purpose of this patch is to introduce plugin aware attributes -
mostly counts of processed jobs.

Plugins cover all the offerings of Oban 2.15 and Oban Pro 1.0, the
current versions at time of writing.

While testing this in a real app, I've noticed issues with exception
handlers for plugin. Decided to slip in this patch changes to both job
and plugin event handlers to rely on standard :telemetry.span/3
attributes instead of :error key (which seems like an artifact of
earlier versions).